move inner classes to another class
